I have done ProductController, besides Keycloak + choose currency

+++ user-service/src/main/java/com/example/repository/ProductRepository.java
@@ -6,6 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -24,7 +26,7 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
 
     @Modifying
     @Query("UPDATE Product p SET p.price = :price WHERE p.id = :id")
-    void updatePriceById(Long id, double price);
+    void updatePriceById(Long id, BigDecimal price);
 
     @Modifying
     @Query("UPDATE Product p SET p.tag = :tag WHERE p.id = :id")
